requirements.txt
python-dotenv-1.0.1
Flask==3.1.0
flask-wtf==1.2.2
email-validator==2.2.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-login==0.6.3
flask-session==0.8.0
pyjwt==2.10.0
flask-mail==0.10.0